Call static `Q*Application.exec` when necessary
