+++ qtpy/QtWidgets.py
@@ -9,7 +9,7 @@
 """Provides widget classes and functions."""
 
 from . import PYQT5, PYQT6, PYSIDE2, PYSIDE6, QtModuleNotInstalledError
-from .utils import getattr_missing_optional_dep
+from .utils import _possibly_static_exec, getattr_missing_optional_dep
 
 
 _missing_optional_names = {}
@@ -19,17 +19,6 @@ def __getattr__(name):
     """Custom getattr to chain and wrap errors due to missing optional deps."""
     raise getattr_missing_optional_dep(
         name, module_name=__name__, optional_names=_missing_optional_names)
-
-
-def _possibly_static_exec(cls, *args, **kwargs):
-    """Call `self.exec` when `self` is given or a static method otherwise."""
-    if isinstance(args[0], cls):
-        if len(args) == 1 and not kwargs:
-            # A special case to avoid the function resolving error
-            return args[0].exec()
-        return args[0].exec(*args[1:], **kwargs)
-    else:
-        return cls.exec(*args, **kwargs)
 
 
 if PYQT5:
@@ -58,7 +47,7 @@ elif PYQT6:
     QPlainTextEdit.setTabStopWidth = lambda self, *args, **kwargs: self.setTabStopDistance(*args, **kwargs)
     QPlainTextEdit.tabStopWidth = lambda self, *args, **kwargs: self.tabStopDistance(*args, **kwargs)
     QPlainTextEdit.print_ = lambda self, *args, **kwargs: self.print(*args, **kwargs)
-    QApplication.exec_ = QApplication.exec
+    QApplication.exec_ = lambda *args, **kwargs: _possibly_static_exec(QApplication, *args, **kwargs)
     QDialog.exec_ = lambda self, *args, **kwargs: self.exec(*args, **kwargs)
     QMenu.exec_ = lambda *args, **kwargs: _possibly_static_exec(QMenu, *args, **kwargs)
     QLineEdit.getTextMargins = lambda self: (self.textMargins().left(), self.textMargins().top(), self.textMargins().right(), self.textMargins().bottom())
@@ -93,6 +82,6 @@ elif PYSIDE6:
     QLineEdit.getTextMargins = lambda self: (self.textMargins().left(), self.textMargins().top(), self.textMargins().right(), self.textMargins().bottom())
 
     # Map DeprecationWarning methods
-    QApplication.exec_ = QApplication.exec
+    QApplication.exec_ = lambda *args, **kwargs: _possibly_static_exec(QApplication, *args, **kwargs)
     QDialog.exec_ = lambda self, *args, **kwargs: self.exec(*args, **kwargs)
     QMenu.exec_ = lambda *args, **kwargs: _possibly_static_exec(QMenu, *args, **kwargs)
